pyproject.toml
[build-system]
requires = ["hatchling>=1.18.0"]
build-backend = "hatchling.build"

[project]
name = "glean_agent_toolkit"
version = "0.1.0"
description = "Universal Tool/Action SDK for Glean agent frameworks"
readme = "README.md"
authors = [{ name = "Glean", email = "dev@glean.com" }]
license = {text = "MIT"}
requires-python = ">=3.10,<4.0"
dependencies = [
    "pydantic>=2.7",
    "typer>=0.12.0",
    "rich>=13.0.0",
    "shellingham>=1.5.0"
]

[project.optional-dependencies]
dev = [
    "commitizen>=4.4.1",
]
test = [
    "pytest>=7.4.3",
    "pytest-asyncio>=0.23.2",
    "pytest-cov>=4.1.0",
]
openai = [
    "openai>=1.0.0",
]
adk = [
    "google-generativeai>=0.4.0",
]
langchain = [
    "langchain>=0.1.0",
]
crewai = [
    "crewai>=0.28.0",
]
codespell = [
    "codespell>=2.2.6",
]
lint = [
    "ruff>=0.5",
]
typing = [
    "mypy>=1.10",
]

[project.scripts]
glean-toolkit = "toolkit.cli:app"

[project.urls]
"Source Code" = "https://github.com/gleanwork/glean-agent-toolkit"

[tool.hatch.build.targets.wheel]
packages = ["glean_agent_toolkit"]

[tool.mypy]
disallow_untyped_defs = "True"
exclude = ["docs/", "tests/"]

[tool.ruff]
line-length = 100
target-version = "py310"
exclude = [
    "docs/",
    ".venv/",
    "**/site-packages/",
    "__pycache__/",
    "dist/",
    "build/"
]

[tool.ruff.lint]
select = ["E", "F", "I", "W", "N", "D", "UP", "ANN", "T201"]
ignore = ["ANN101", "D203", "D212"]

[tool.ruff.lint.isort]
known-first-party = ["glean_agent_toolkit", "toolkit"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.coverage.run]
omit = ["tests/*"]

[tool.pytest.ini_options]
addopts = "--strict-markers --strict-config --durations=5"
testpaths = ["tests"]
python_files = "test_*.py"
asyncio_mode = "auto" 